Matrix Computation
This version changes completely the way the integral is computed.
It just uses two for-loops to run over the output screen. In each pixel of the output screen, two matrices are computed, the kernel matrix and the obliquity matrix, so that the computation of the intensity in the local output pixel is carried by matrix operations, saving two for loops compared with the previous version